+++ CHANGES.md
@@ -1,3 +1,8 @@
+## unreleased
+
+- add a `swhid` executable able to download any kind of identifiers and to compute some of them
+- add a `Compute.content_identifier_from_file` function
+
 ## 0.2 - 2022-06-21
 
 - move all the things without dependency to `swhid_core`